codex: address PR review feedback (#25)
