+++ test/tools.test.ts
@@ -1094,19 +1094,7 @@ describe("webSearch tool", () => {
       }),
     });
 
-  test("uses custom web search tool for openai provider", async () => {
-    const dir = await tmpDir();
-    const t: any = createWebSearchTool(
-      makeCtx(dir, {
-        config: makeConfig(dir, { provider: "openai", model: "gpt-5.2", subAgentModel: "gpt-5.2" }),
-      })
-    );
-    expect(t.type).toBeUndefined();
-    expect(typeof t.execute).toBe("function");
-    expect(t.description).toContain("BRAVE_API_KEY");
-  });
-
-  test("uses Exa-backed web search for google provider (Gemini model IDs)", async () => {
+  test("uses Exa-backed web search", async () => {
     const dir = await tmpDir();
     const t: any = createWebSearchTool(
       makeCtx(dir, {
@@ -1120,31 +1108,11 @@ describe("webSearch tool", () => {
     expect(t.type).toBeUndefined();
     expect(typeof t.execute).toBe("function");
     expect(t.description).toContain("EXA_API_KEY");
-    expect(t.description).not.toContain("BRAVE_API_KEY");
   });
 
-  test("uses Exa-backed web search for google provider (non-Gemini model IDs)", async () => {
+  test("web search requires EXA_API_KEY", async () => {
     const dir = await tmpDir();
-    const t: any = createWebSearchTool(
-      makeCtx(dir, {
-        config: makeConfig(dir, {
-          provider: "google",
-          model: "text-bison",
-          subAgentModel: "text-bison",
-        }),
-      })
-    );
-    expect(t.type).toBeUndefined();
-    expect(typeof t.execute).toBe("function");
-    expect(t.description).toContain("EXA_API_KEY");
-    expect(t.description).not.toContain("BRAVE_API_KEY");
-  });
-
-  test("google provider web search requires EXA_API_KEY", async () => {
-    const dir = await tmpDir();
-    const oldBrave = process.env.BRAVE_API_KEY;
     const oldExa = process.env.EXA_API_KEY;
-    process.env.BRAVE_API_KEY = "test-brave-key";
     delete process.env.EXA_API_KEY;
 
     try {
@@ -1160,90 +1128,14 @@ describe("webSearch tool", () => {
       const out: string = await t.execute({ query: "test", maxResults: 1 });
       expect(out).toContain("set EXA_API_KEY");
     } finally {
-      if (oldBrave) process.env.BRAVE_API_KEY = oldBrave;
-      else delete process.env.BRAVE_API_KEY;
-      if (oldExa) process.env.EXA_API_KEY = oldExa;
-    }
-  });
-
-  test("google provider web search uses saved Exa API key when env key is missing", async () => {
-    const dir = await tmpDir();
-    const paths = getAiCoworkerPaths({ homedir: dir });
-    await writeConnectionStore(paths, {
-      version: 1,
-      updatedAt: new Date().toISOString(),
-      services: {},
-      toolApiKeys: {
-        exa: "saved-exa-key",
-      },
-    });
-
-    const oldBrave = process.env.BRAVE_API_KEY;
-    const oldExa = process.env.EXA_API_KEY;
-    delete process.env.BRAVE_API_KEY;
-    delete process.env.EXA_API_KEY;
-
-    const originalFetch = globalThis.fetch;
-    globalThis.fetch = mock(async (_url: any, init: any) => {
-      expect(init?.headers?.["x-api-key"]).toBe("saved-exa-key");
-      return new Response(
-        JSON.stringify({
-          results: [
-            {
-              title: "Stored Exa Result",
-              url: "https://exa.example",
-              text: { text: "Stored key content" },
-            },
-          ],
-        }),
-        { status: 200, headers: { "Content-Type": "application/json" } }
-      );
-    }) as any;
-
-    try {
-      const t: any = createWebSearchTool(
-        makeCtx(dir, {
-          config: makeConfig(dir, {
-            provider: "google",
-            model: "gemini-3.1-pro-preview",
-            subAgentModel: "gemini-3.1-pro-preview",
-          }),
-        })
-      );
-      const out: string = await t.execute({ query: "test query", maxResults: 3 });
-      expect(out).toContain("Stored Exa Result");
-      expect(out).toContain("Stored key content");
-    } finally {
-      globalThis.fetch = originalFetch;
-      if (oldBrave) process.env.BRAVE_API_KEY = oldBrave;
-      else delete process.env.BRAVE_API_KEY;
       if (oldExa) process.env.EXA_API_KEY = oldExa;
       else delete process.env.EXA_API_KEY;
     }
   });
 
-  test("uses custom web search tool for anthropic provider", async () => {
-    const dir = await tmpDir();
-    const t: any = createWebSearchTool(
-      makeCtx(dir, {
-        config: makeConfig(dir, {
-          provider: "anthropic",
-          model: "claude-opus-4-6",
-          subAgentModel: "claude-opus-4-6",
-        }),
-      })
-    );
-    expect(t.type).toBeUndefined();
-    expect(typeof t.execute).toBe("function");
-    expect(t.description).toContain("BRAVE_API_KEY");
-  });
-
   test("returns disabled message without API keys", async () => {
     const dir = await tmpDir();
-
-    const oldBrave = process.env.BRAVE_API_KEY;
     const oldExa = process.env.EXA_API_KEY;
-    delete process.env.BRAVE_API_KEY;
     delete process.env.EXA_API_KEY;
 
     try {
@@ -1251,377 +1143,6 @@ describe("webSearch tool", () => {
       const out: string = await t.execute({ query: "test", maxResults: 1 });
       expect(out).toContain("webSearch disabled");
     } finally {
-      if (oldBrave) process.env.BRAVE_API_KEY = oldBrave;
-      if (oldExa) process.env.EXA_API_KEY = oldExa;
-    }
-  });
-
-  test("returns a guidance message for empty or whitespace-only query", async () => {
-    const dir = await tmpDir();
-    const t: any = createWebSearchTool(makeCustomSearchCtx(dir));
-    const out: string = await t.execute({ query: "   ", maxResults: 1 });
-    expect(out).toContain("non-empty query");
-  });
-
-  test("falls back to turnUserPrompt when query is missing", async () => {
-    const dir = await tmpDir();
-
-    const oldExa = process.env.EXA_API_KEY;
-    process.env.EXA_API_KEY = "test-exa-key";
-
-    const originalFetch = globalThis.fetch;
-    globalThis.fetch = mock(async (url: any, init: any) => {
-      expect(String(url)).toContain("api.exa.ai/search");
-      const body = JSON.parse(String(init?.body ?? "{}"));
-      expect(body.query).toBe("latest Apple earnings guidance");
-      expect(body.numResults).toBe(2);
-      return new Response(
-        JSON.stringify({
-          results: [
-            {
-              title: "Fallback Result",
-              url: "https://example.com/fallback",
-              text: "Fallback content",
-            },
-          ],
-        }),
-        { status: 200, headers: { "Content-Type": "application/json" } }
-      );
-    }) as any;
-
-    try {
-      const t: any = createWebSearchTool(
-        makeCtx(dir, {
-          config: makeConfig(dir, {
-            provider: "google",
-            model: "gemini-3.1-pro-preview",
-            subAgentModel: "gemini-3.1-pro-preview",
-          }),
-          turnUserPrompt: "latest Apple earnings guidance",
-        })
-      );
-
-      const out: string = await t.execute({
-        maxResults: 2,
-      });
-      expect(out).toContain("Fallback Result");
-      expect((globalThis.fetch as any).mock.calls.length).toBe(1);
-    } finally {
-      globalThis.fetch = originalFetch;
-      if (oldExa) process.env.EXA_API_KEY = oldExa;
-      else delete process.env.EXA_API_KEY;
-    }
-  });
-
-  test("accepts compatibility query aliases and provider-native extra keys", async () => {
-    const dir = await tmpDir();
-
-    const oldBrave = process.env.BRAVE_API_KEY;
-    const oldExa = process.env.EXA_API_KEY;
-    process.env.BRAVE_API_KEY = "test-brave-key";
-    delete process.env.EXA_API_KEY;
-
-    const seenQueries: string[] = [];
-    const originalFetch = globalThis.fetch;
-    globalThis.fetch = mock(async (url: any) => {
-      const query = new URL(String(url)).searchParams.get("q") ?? "";
-      seenQueries.push(query);
-      return new Response(
-        JSON.stringify({
-          web: {
-            results: [
-              {
-                title: `Result for ${query}`,
-                url: "https://example.com",
-                description: "Alias compatibility",
-              },
-            ],
-          },
-        }),
-        { status: 200, headers: { "Content-Type": "application/json" } }
-      );
-    }) as any;
-
-    try {
-      const t: any = createWebSearchTool(makeCustomSearchCtx(dir));
-      const aliasInputs = [
-        { q: "q alias query" },
-        { searchQuery: "searchQuery alias query" },
-        { text: "text alias query" },
-        { prompt: "prompt alias query" },
-      ] as const;
-
-      for (const input of aliasInputs) {
-        const query = Object.values(input)[0];
-        const out: string = await t.execute({
-          ...input,
-          mode: "auto",
-          dynamicThreshold: 0.42,
-          maxResults: 1,
-        });
-        expect(out).toContain(`Result for ${query}`);
-      }
-
-      expect(seenQueries).toEqual([
-        "q alias query",
-        "searchQuery alias query",
-        "text alias query",
-        "prompt alias query",
-      ]);
-    } finally {
-      globalThis.fetch = originalFetch;
-      if (oldBrave) process.env.BRAVE_API_KEY = oldBrave;
-      else delete process.env.BRAVE_API_KEY;
-      if (oldExa) process.env.EXA_API_KEY = oldExa;
-      else delete process.env.EXA_API_KEY;
-    }
-  });
-
-  test("accepts unknown compatibility fields when query is present", async () => {
-    const dir = await tmpDir();
-
-    const oldBrave = process.env.BRAVE_API_KEY;
-    const oldExa = process.env.EXA_API_KEY;
-    process.env.BRAVE_API_KEY = "test-brave-key";
-    delete process.env.EXA_API_KEY;
-
-    const originalFetch = globalThis.fetch;
-    globalThis.fetch = mock(async (url: any) => {
-      const query = new URL(String(url)).searchParams.get("q") ?? "";
-      expect(query).toBe("query with unknown extras");
-      return new Response(
-        JSON.stringify({
-          web: {
-            results: [
-              {
-                title: "Compatibility Result",
-                url: "https://example.com/compat",
-                description: "Unknown key compatibility",
-              },
-            ],
-          },
-        }),
-        { status: 200, headers: { "Content-Type": "application/json" } }
-      );
-    }) as any;
-
-    try {
-      const t: any = createWebSearchTool(makeCustomSearchCtx(dir));
-      const out: string = await t.execute({
-        query: "query with unknown extras",
-        search_context_size: "low",
-        user_location: {
-          type: "approximate",
-          city: "Austin",
-        },
-        include_domains: ["example.com"],
-        maxResults: 1,
-      } as any);
-
-      expect(out).toContain("Compatibility Result");
-      expect(out).not.toContain("requires a query");
-    } finally {
-      globalThis.fetch = originalFetch;
-      if (oldBrave) process.env.BRAVE_API_KEY = oldBrave;
-      else delete process.env.BRAVE_API_KEY;
-      if (oldExa) process.env.EXA_API_KEY = oldExa;
-      else delete process.env.EXA_API_KEY;
-    }
-  });
-
-  test("uses BRAVE_API_KEY when available", async () => {
-    const dir = await tmpDir();
-
-    const oldBrave = process.env.BRAVE_API_KEY;
-    const oldExa = process.env.EXA_API_KEY;
-    process.env.BRAVE_API_KEY = "test-brave-key";
-    delete process.env.EXA_API_KEY;
-
-    const originalFetch = globalThis.fetch;
-    globalThis.fetch = mock(async (url: any) => {
-      return new Response(
-        JSON.stringify({
-          web: {
-            results: [
-              {
-                title: "Test Result",
-                url: "https://example.com",
-                description: "A test description",
-              },
-            ],
-          },
-        }),
-        { status: 200, headers: { "Content-Type": "application/json" } }
-      );
-    }) as any;
-
-    try {
-      const t: any = createWebSearchTool(makeCustomSearchCtx(dir));
-      const out: string = await t.execute({ query: "test query", maxResults: 5 });
-      expect(out).toContain("Test Result");
-      expect(out).toContain("https://example.com");
-      expect(out).toContain("A test description");
-    } finally {
-      globalThis.fetch = originalFetch;
-      if (oldBrave) process.env.BRAVE_API_KEY = oldBrave;
-      else delete process.env.BRAVE_API_KEY;
-      if (oldExa) process.env.EXA_API_KEY = oldExa;
-    }
-  });
-
-  test("uses EXA_API_KEY as fallback", async () => {
-    const dir = await tmpDir();
-
-    const oldBrave = process.env.BRAVE_API_KEY;
-    const oldExa = process.env.EXA_API_KEY;
-    delete process.env.BRAVE_API_KEY;
-    process.env.EXA_API_KEY = "test-exa-key";
-
-    const originalFetch = globalThis.fetch;
-    globalThis.fetch = mock(async (url: any) => {
-      expect(String(url)).toContain("api.exa.ai/search");
-      return new Response(
-        JSON.stringify({
-          results: [
-            {
-              title: "Exa Result",
-              url: "https://exa.com",
-              text: { text: "Exa content" },
-            },
-          ],
-        }),
-        { status: 200, headers: { "Content-Type": "application/json" } }
-      );
-    }) as any;
-
-    try {
-      const t: any = createWebSearchTool(makeCustomSearchCtx(dir));
-      const out: string = await t.execute({ query: "test query", maxResults: 5 });
-      expect(out).toContain("Exa Result");
-      expect(out).toContain("https://exa.com");
-      expect(out).toContain("Exa content");
-    } finally {
-      globalThis.fetch = originalFetch;
-      if (oldBrave) process.env.BRAVE_API_KEY = oldBrave;
-      if (oldExa) process.env.EXA_API_KEY = oldExa;
-      else delete process.env.EXA_API_KEY;
-    }
-  });
-
-  test("handles Brave API error response", async () => {
-    const dir = await tmpDir();
-
-    const oldBrave = process.env.BRAVE_API_KEY;
-    const oldExa = process.env.EXA_API_KEY;
-    process.env.BRAVE_API_KEY = "test-brave-key";
-    delete process.env.EXA_API_KEY;
-
-    const originalFetch = globalThis.fetch;
-    globalThis.fetch = mock(async () => {
-      return new Response("Unauthorized", { status: 401, statusText: "Unauthorized" });
-    }) as any;
-
-    try {
-      const t: any = createWebSearchTool(makeCustomSearchCtx(dir));
-      const out: string = await t.execute({ query: "test", maxResults: 1 });
-      expect(out).toContain("Brave search failed");
-      expect(out).toContain("401");
-    } finally {
-      globalThis.fetch = originalFetch;
-      if (oldBrave) process.env.BRAVE_API_KEY = oldBrave;
-      else delete process.env.BRAVE_API_KEY;
-      if (oldExa) process.env.EXA_API_KEY = oldExa;
-    }
-  });
-
-  test("handles Exa API error response", async () => {
-    const dir = await tmpDir();
-
-    const oldBrave = process.env.BRAVE_API_KEY;
-    const oldExa = process.env.EXA_API_KEY;
-    delete process.env.BRAVE_API_KEY;
-    process.env.EXA_API_KEY = "test-exa-key";
-
-    const originalFetch = globalThis.fetch;
-    globalThis.fetch = mock(async () => {
-      return new Response(JSON.stringify({ message: "Internal Server Error" }), {
-        status: 500,
-        statusText: "Internal Server Error",
-      });
-    }) as any;
-
-    try {
-      const t: any = createWebSearchTool(makeCustomSearchCtx(dir));
-      const out: string = await t.execute({ query: "test", maxResults: 1 });
-      expect(out).toContain("Exa search failed");
-      expect(out).toContain("Internal Server Error");
-    } finally {
-      globalThis.fetch = originalFetch;
-      if (oldBrave) process.env.BRAVE_API_KEY = oldBrave;
-      if (oldExa) process.env.EXA_API_KEY = oldExa;
-      else delete process.env.EXA_API_KEY;
-    }
-  });
-
-  test("handles empty results from Brave", async () => {
-    const dir = await tmpDir();
-
-    const oldBrave = process.env.BRAVE_API_KEY;
-    const oldExa = process.env.EXA_API_KEY;
-    process.env.BRAVE_API_KEY = "test-brave-key";
-    delete process.env.EXA_API_KEY;
-
-    const originalFetch = globalThis.fetch;
-    globalThis.fetch = mock(async () => {
-      return new Response(JSON.stringify({ web: { results: [] } }), {
-        status: 200,
-        headers: { "Content-Type": "application/json" },
-      });
-    }) as any;
-
-    try {
-      const t: any = createWebSearchTool(makeCustomSearchCtx(dir));
-      const out: string = await t.execute({ query: "nothing", maxResults: 5 });
-      expect(out).toBe("No results");
-    } finally {
-      globalThis.fetch = originalFetch;
-      if (oldBrave) process.env.BRAVE_API_KEY = oldBrave;
-      else delete process.env.BRAVE_API_KEY;
-      if (oldExa) process.env.EXA_API_KEY = oldExa;
-    }
-  });
-
-  test("Brave is preferred over Exa when both keys exist", async () => {
-    const dir = await tmpDir();
-
-    const oldBrave = process.env.BRAVE_API_KEY;
-    const oldExa = process.env.EXA_API_KEY;
-    process.env.BRAVE_API_KEY = "test-brave-key";
-    process.env.EXA_API_KEY = "test-exa-key";
-
-    const originalFetch = globalThis.fetch;
-    let calledUrl = "";
-    globalThis.fetch = mock(async (url: any) => {
-      calledUrl = String(url);
-      return new Response(
-        JSON.stringify({
-          web: {
-            results: [{ title: "Brave Hit", url: "https://brave.com", description: "desc" }],
-          },
-        }),
-        { status: 200, headers: { "Content-Type": "application/json" } }
-      );
-    }) as any;
-
-    try {
-      const t: any = createWebSearchTool(makeCustomSearchCtx(dir));
-      const out: string = await t.execute({ query: "test", maxResults: 1 });
-      expect(calledUrl).toContain("brave");
-      expect(out).toContain("Brave Hit");
-    } finally {
-      globalThis.fetch = originalFetch;
-      if (oldBrave) process.env.BRAVE_API_KEY = oldBrave;
-      else delete process.env.BRAVE_API_KEY;
       if (oldExa) process.env.EXA_API_KEY = oldExa;
       else delete process.env.EXA_API_KEY;
     }
